backend-engineer/microservices/user_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
PyJWT
redis
python-dotenv
pydantic[email] 